more stubbing done, and worked on render setup system
created BindMgr Stubs

created ability to switch between Vulkan and OpenGL Render
and started stubbing the OpenGL Render.

added an Abstract SysRender class to have independent render for
OGL and VK and RenderStruct will be a wrapper forwarding all calls
to the current SysRender stored in the private part of RenderStruct.