plenty of dirty hacks in here..
fighting to manage the compilation of the horrible windows ridden
code, with crazy defines/typedefs. There's no guarantie that this
will work on any level... only hope that an so file of sorts comes
out on the other side.

clientRes is completely empty... will need to find a way to
"compile" in all the resource strings in to a loadable library.